05092018 seeders

+++ database/migrations/2018_07_11_003807_create_reciclators_table.php
@@ -21,7 +21,7 @@ class CreateReciclatorsTable extends Migration
             $table->string('last_name');
             $table->string('address');
             $table->integer('phone_number');
-            $table->integer('mobile_number');
+            $table->bigInteger('mobile_number');
             $table->unsignedInteger('id_type_purchase')->comment('Id de tipo de compra de la tabla tipos de compra');            
             $table->timestamps();
             $table->engine    = 'MyIsam';
